changes to make it easier to use just purrr::map or furrr::future_map
Initially I was using map_at, and this made it hard to remember if I wanted to do parallel on a single file. It was originally for choosing if you wanted assigned data only or not. So, now we have the option of returning assigned only, and we just filter the peaks into the data extraction, with the simplest case being done first so it is faster.