function and workflows to work with the SEMFs that share peaks
The logic for this is to find those SEMFs that share peaks, extract them,
and then start checking for those that have > 50% of peaks from one of the
SEMFs. If they share > 50% of peaks, they get merged together.
After merging, go through and do voting again, with matching by
M/Z to see if we pick up peaks from some of the other GEMFs